Solution to the problem with loss of a capture of a very fast dragging window.
When we drag a window very fast, we can lose the capture of mouse operations.

The problem in time between SendLeftMouseButtonUp and SendLeftMouseButtonDown operations.

I not see any reason for calling SendLeftMouseButtonUp. This operation can lead to lose capture of mouse events.

By fact we need to just notify a new window about MouseDown and let it move.